Challenge-2/project/requirements.txt
Django==5.0.2
djangorestframework==3.15.2
django-filter==24.3
django-environ==0.11.2
google-generativeai
ipython~=8.20.0
